[refactor] 빌드 가능
